Revert "Addressing comments"
This reverts commit 45d5830e16939360293512e6f41f13d364ce06ae.
