+++ py-utils/test/cli_framework/cli_test_data/users_show_table_output
